src/@core/data/usecases/authentication/remote-authentication.ts
import { HttpClient, HttpStatusCode } from '@/@core/data/protocols/http'
import { InvalidCredentialsError, UnexpectedError } from '@/@core/domain/errors'
import { Authentication } from '@/@core/domain/usecases'

export class RemoteAuthentication implements Authentication {
  constructor(
    private readonly url: string,
    private readonly httpClient: HttpClient<Authentication.Model>
  ) {}

  async auth(params: Authentication.Params): Promise<Authentication.Model> {
    const httpResponse = await this.httpClient.request({
      url: this.url,
      method: 'post',
      body: params
    })
    switch (httpResponse.statusCode) {
      case HttpStatusCode.ok:
        return httpResponse.body
      case HttpStatusCode.unauthorized:
        throw new InvalidCredentialsError()
      default:
        throw new UnexpectedError()
    }
  }
}
